Rename `git_hub` to `github`
GitHub is a single word, as represented by github.com.
