+++ lib/rubocop/formatter/github_actions_formatter.rb
