+++ spec/rubocop/formatter/github_actions_formatter_spec.rb
